#2 Refactored hw01.
