+++ kuklinVD/src/main/java/com/cbs/edu/hw01/Sprint.java
@@ -4,5 +4,9 @@ package com.cbs.edu.hw01;
  * Parts of the project 22.02.17.
  */
 public class Sprint {
-    String endPhaseDate;
+    String title;
+    String startDate;
+    String endDate;
+    Project project;
+    Task[] tasks;
 }
