+++ kuklinVD/src/main/java/com/cbs/edu/hw01/Task.java
@@ -4,5 +4,9 @@ package com.cbs.edu.hw01;
  * Parts of the sprint 22.02.17.
  */
 public class Task {
+    String title;
+    String startDate;
+    String endDate;
     int runTime;
+    Task[] subTasks;
 }
